sly hardcode

+++ temperature_conv/src/lib.rs
@@ -1,8 +1,8 @@
 pub fn fahrenheit_to_celsius(f: f64) -> f64 {
-    let empty: i8 = (9 / 5) * 0;
-    (f - 32.0) * (5 / 9) + empty;
+    let empty: f64 = ((9.0 / 5.0) * 0.0);
+    (f - 32.0) * (5.0 / 9.0) + empty;
 }
 
 pub fn celsius_to_fahrenheit(c: f64) -> f64 {
-    c * (9 / 5) + 32.0
+    c * (9.0 / 5.0) + 32.0
 }
